+++ src/components/info-panel/InfoPanel.tsx
@@ -1,11 +1,8 @@
 /** @jsxImportSource @emotion/react */
-import { InputGroup } from '@blueprintjs/core';
+import { Icon, InputGroup } from '@blueprintjs/core';
 import { css } from '@emotion/react';
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment, @typescript-eslint/prefer-ts-expect-error
-// @ts-ignore
 import { Disclosure } from '@headlessui/react';
-import { CSSProperties, useState } from 'react';
-import { FaChevronRight } from 'react-icons/fa';
+import { CSSProperties, useCallback, useMemo, useState } from 'react';
 
 import { ValueRenderers } from '../index';
 import { Table } from '../table/Table';
@@ -31,87 +28,165 @@ const style = {
       transition: 'all 0.3s ease-in-out',
     }),
   button: css({
+    borderBottom: '1px solid #f5f5f5',
     display: 'flex',
     alignItems: 'center',
     gap: 5,
     padding: '5px 2px',
+    width: '100%',
+    ':hover': {
+      backgroundColor: '#f5f5f5',
+    },
   }),
 };
 
 export function InfoPanel(props: InfoPanelProps) {
   const [search, setSearch] = useState('');
   const { title = 'Information', data = [], titleStyle, inputStyle } = props;
-  function viewData(data: Record<string, string | number | object>) {
-    const exactMatch: Array<[string, string | number | object]> = [];
-    const startsWith: Array<[string, string | number | object]> = [];
-    const includes: Array<[string, string | number | object]> = [];
-    const valueContains: Array<[string, string | number | object]> = [];
+  const viewData = useCallback(
+    (data: Record<string, string | number | object>) => {
+      const exactMatch: Array<[string, string | number | object]> = [];
+      const startsWith: Array<[string, string | number | object]> = [];
+      const includes: Array<[string, string | number | object]> = [];
+      const valueContains: Array<[string, string | number | object]> = [];
 
-    for (const [key, value] of Object.entries(data)) {
-      if (key === search) {
-        exactMatch.push([key, value]);
-        continue;
+      for (const [key, value] of Object.entries(data).sort(([a], [b]) =>
+        a.localeCompare(b),
+      )) {
+        const lowerKey = key.toLowerCase();
+        const lowerSearch = search.toLowerCase();
+        if (lowerKey === lowerSearch) {
+          exactMatch.push([key, value]);
+          continue;
+        }
+        if (lowerKey.startsWith(lowerSearch)) {
+          startsWith.push([key, value]);
+          continue;
+        }
+        if (lowerKey.includes(lowerSearch)) {
+          includes.push([key, value]);
+          continue;
+        }
+        if (valueSearch(value, search)) {
+          valueContains.push([key, value]);
+          continue;
+        }
       }
-      if (key.startsWith(search)) {
-        startsWith.push([key, value]);
-        continue;
+      return [...exactMatch, ...startsWith, ...includes, ...valueContains];
+    },
+    [search],
+  );
+  const { filteredData, total, count } = useMemo(() => {
+    const filteredData: Array<
+      Omit<InfoPanelData, 'data'> & {
+        data: Array<[string, string | number | object]>;
       }
-      if (key.includes(search)) {
-        includes.push([key, value]);
-        continue;
-      }
-      if (valueSearch(value, search)) {
-        valueContains.push([key, value]);
-        continue;
-      }
+    > = [];
+    let total = 0;
+    let count = 0;
+    for (const { data: dataContent, ...other } of data) {
+      total += Object.entries(dataContent).length;
+      const content = viewData(dataContent);
+      if (content.length === 0) continue;
+      filteredData.push({ data: content, ...other });
+      count += content.length;
     }
-    return [...exactMatch, ...startsWith, ...includes, ...valueContains].map(
-      ([key, value]) => (
-        <Table.Row key={key}>
-          <ValueRenderers.Text value={key} />
-          {valueCell(value)}
-        </Table.Row>
-      ),
-    );
-  }
+    return { filteredData, total, count };
+  }, [data, viewData]);
   return (
     <div css={style.container}>
       <div style={titleStyle}>{title}</div>
-      <InputGroup
-        placeholder="search for a parameter ..."
-        value={search}
-        onChange={({ target }) => {
-          if (target.value !== undefined) setSearch(target.value);
+      <div
+        tabIndex={0}
+        css={css({
+          zIndex: 10,
+          marginTop: '5px',
+          position: 'sticky',
+          backgroundColor: 'white',
+          top: '5px',
+          display: 'flex',
+          alignItems: 'center',
+          gap: 6,
+          width: '100%',
+        })}
+      >
+        <InputGroup
+          css={css({
+            flexGrow: 1,
+          })}
+          placeholder="search for a parameter ..."
+          value={search}
+          onChange={({ target }) => {
+            if (target.value !== undefined) {
+              setSearch(target.value);
+            }
+          }}
+          style={inputStyle}
+          leftIcon="search"
+          type="search"
+          fill
+        />
+        [{count}/{total}]
+      </div>
+      <div
+        style={{
+          marginTop: '5px',
+          display: 'flex',
+          flexDirection: 'column',
+          gap: '5px',
         }}
-        style={inputStyle}
-        leftIcon="search"
-        type="search"
-      />
-
-      {data.map(({ description, data }) => {
-        const content = viewData(data);
-        return content.length > 0 ? (
-          <Disclosure defaultOpen key={description}>
-            {({ open }) => (
-              <>
-                <Disclosure.Button css={style.button}>
-                  <FaChevronRight css={style.chevron(open)} />
-                  {description}
-                </Disclosure.Button>
-                <Disclosure.Panel>
-                  <Table bordered>
-                    <Table.Header>
-                      <ValueRenderers.Header value="Parameter" />
-                      <ValueRenderers.Header value="Value" />
-                    </Table.Header>
-                    {content}
-                  </Table>
-                </Disclosure.Panel>
-              </>
-            )}
-          </Disclosure>
-        ) : null;
-      })}
+      >
+        {filteredData.map(({ description, data }) => {
+          return (
+            <Disclosure defaultOpen key={description}>
+              {({ open }) => (
+                <>
+                  <Disclosure.Button css={style.button}>
+                    <Icon icon="chevron-right" css={style.chevron(open)} />
+                    {description}
+                  </Disclosure.Button>
+                  <Disclosure.Panel>
+                    <Table
+                      striped
+                      css={css({
+                        width: '100%',
+                        'td:first-child': {
+                          width: '30%',
+                        },
+                      })}
+                      compact
+                    >
+                      <Table.Header>
+                        <ValueRenderers.Header
+                          style={{ width: '50px' }}
+                          value="Parameter"
+                        />
+                        <ValueRenderers.Header value="Value" />
+                      </Table.Header>
+                      {data.map(([key, value]) => (
+                        <Table.Row
+                          key={key}
+                          bordered
+                          style={{
+                            height: '10px',
+                            padding: '0 !imporant',
+                          }}
+                        >
+                          <ValueRenderers.Text
+                            style={{ width: '50px' }}
+                            value={key}
+                          />
+                          {valueCell(value)}
+                        </Table.Row>
+                      ))}
+                    </Table>
+                  </Disclosure.Panel>
+                </>
+              )}
+            </Disclosure>
+          );
+        })}
+      </div>
     </div>
   );
 }
@@ -145,7 +220,12 @@ function valueCell(value: number | string | object | boolean) {
 function valueSearch(
   value: number | string | object | boolean,
   search: string,
+  lowerCase = true,
 ): boolean {
+  if (lowerCase) {
+    value = String(value).toLowerCase();
+    search = search.toLowerCase();
+  }
   switch (typeof value) {
     case 'number':
       return String(value).includes(search);