+++ stories/components/toolbar.stories.tsx
@@ -1,6 +1,6 @@
 import { Menu, MenuItem } from '@blueprintjs/core';
 import { Meta } from '@storybook/react';
-import { JSX, useState } from 'react';
+import { CSSProperties, JSX, useState } from 'react';
 import { BiClipboard, BiCreditCard, BiPaperclip } from 'react-icons/bi';
 import { FaClipboard, FaCreditCard, FaPaperclip } from 'react-icons/fa6';
 import { HiClipboard, HiCreditCard, HiOutlinePaperClip } from 'react-icons/hi2';
@@ -26,89 +26,89 @@ export default {
 } as Meta<ToolbarProps>;
 
 type ToolbarItems = Array<
-  Pick<ToolbarItemProps, 'title' | 'icon' | 'disabled'> & {
+  Pick<ToolbarItemProps, 'tooltip' | 'icon' | 'disabled'> & {
     id: string;
     content?: JSX.Element;
   }
 >;
 
 const itemsBlueprintIcons: ToolbarItems = [
-  { id: 'copy', icon: 'phone', title: 'Phone' },
-  { id: 'paste', icon: 'add-column-left', title: 'Add left' },
-  { id: 'undo', icon: 'add-column-right', title: 'Right' },
+  { id: 'copy', icon: 'phone', tooltip: 'Phone' },
+  { id: 'paste', icon: 'add-column-left', tooltip: 'Add left' },
+  { id: 'undo', icon: 'add-column-right', tooltip: 'Right' },
   {
     id: 'redo',
     icon: 'redo',
-    title: 'Redo',
+    tooltip: 'Redo',
   },
-  { id: 'test1', icon: 'undo', title: 'Undo' },
-  { id: 'test2', icon: 'paperclip', title: 'Attachment' },
-  { id: 'test3', icon: 'help', title: 'Help' },
-  { id: 'test4', icon: 'lab-test', title: 'Lab', disabled: true },
-  { id: 'test5', icon: 'trash', title: 'Trash' },
+  { id: 'test1', icon: 'undo', tooltip: 'Undo' },
+  { id: 'test2', icon: 'paperclip', tooltip: 'Attachment' },
+  { id: 'test3', icon: 'help', tooltip: 'Help' },
+  { id: 'test4', icon: 'lab-test', tooltip: 'Lab', disabled: true },
+  { id: 'test5', icon: 'trash', tooltip: 'Trash' },
 ];
 
 const itemsMixedIcons: ToolbarItems = [
   {
     id: 'paperclip-blueprint',
     icon: 'paperclip',
-    title: 'BlueprintJS paperclip icon',
+    tooltip: 'BlueprintJS paperclip icon',
   },
   {
     id: 'paperclip-fa',
     icon: <FaPaperclip />,
-    title: 'Fontawesome paperclip icon',
+    tooltip: 'Fontawesome paperclip icon',
   },
   {
     id: 'paperclip-hi-2',
     icon: <HiOutlinePaperClip />,
-    title: 'Heroicons 2 paperclip icon',
+    tooltip: 'Heroicons 2 paperclip icon',
   },
   {
     id: 'paperclip-bi',
     icon: <BiPaperclip />,
-    title: 'Box icons paperclip icon',
+    tooltip: 'Box icons paperclip icon',
   },
   {
     id: 'clipboard-blueprint',
     icon: 'clipboard',
-    title: 'BlueprintJS paperclip icon',
+    tooltip: 'BlueprintJS paperclip icon',
   },
   {
     id: 'clipboard-fontawesome',
     icon: <FaClipboard />,
-    title: 'Fontawesome clipboard icon',
+    tooltip: 'Fontawesome clipboard icon',
   },
 
   {
     id: 'clipboard-hi-2',
     icon: <HiClipboard />,
-    title: 'Heroicons 2 clipboard icon',
+    tooltip: 'Heroicons 2 clipboard icon',
   },
   {
     id: 'clipboard-bi',
     icon: <BiClipboard />,
-    title: 'Box icons clipboard icon',
+    tooltip: 'Box icons clipboard icon',
   },
   {
     id: 'credit-card-blueprint',
     icon: 'credit-card',
-    title: 'BlueprintJS credit-card icon',
+    tooltip: 'BlueprintJS credit-card icon',
   },
   {
     id: 'credit-card-fa',
     icon: <FaCreditCard />,
-    title: 'Fontawesome credit card icon',
+    tooltip: 'Fontawesome credit card icon',
   },
   {
     id: 'credit-card-hi2',
     icon: <HiCreditCard />,
-    title: 'Heroicons credit card icon',
+    tooltip: 'Heroicons credit card icon',
   },
   {
     id: 'credit-card-bi',
     icon: <BiCreditCard />,
-    title: 'Box icons credit card icon',
+    tooltip: 'Box icons credit card icon',
     disabled: true,
   },
 ];
@@ -117,7 +117,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'clipboard-bi',
     icon: <BiClipboard />,
-    title: 'Box icons clipboard icon',
+    tooltip: 'Box icons clipboard icon',
     content: (
       <Menu>
         <MenuItem text="Box icons clipboard icon" />
@@ -127,7 +127,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'paste',
     icon: 'add-column-left',
-    title: 'Add left',
+    tooltip: 'Add left',
     content: (
       <Menu>
         <MenuItem text="Add left" />
@@ -137,7 +137,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'paperclip-bi',
     icon: <BiPaperclip />,
-    title: 'Box icons paperclip icon',
+    tooltip: 'Box icons paperclip icon',
     content: (
       <Menu>
         <MenuItem text="Box icons paperclip icon" />
@@ -147,7 +147,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'clipboard-blueprint',
     icon: 'clipboard',
-    title: 'BlueprintJS paperclip icon',
+    tooltip: 'BlueprintJS paperclip icon',
     content: (
       <Menu>
         <MenuItem text="BlueprintJS paperclip icon" />
@@ -157,7 +157,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'credit-card-bi',
     icon: <BiCreditCard />,
-    title: 'Box icons credit card icon',
+    tooltip: 'Box icons credit card icon',
     content: (
       <Menu>
         <MenuItem text="Box icons credit card icon" />
@@ -167,7 +167,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'credit-card-blueprint',
     icon: 'credit-card',
-    title: 'BlueprintJS credit-card icon',
+    tooltip: 'BlueprintJS credit-card icon',
     content: (
       <Menu>
         <MenuItem text="BlueprintJS credit-card icon" />
@@ -177,7 +177,7 @@ const itemsPopover: ToolbarItems = [
   {
     id: 'clipboard-hi-2',
     icon: <HiClipboard />,
-    title: 'Heroicons 2 clipboard icon',
+    tooltip: 'Heroicons 2 clipboard icon',
     content: (
       <Menu>
         <MenuItem text="Heroicons 2 clipboard icon" />
@@ -196,7 +196,7 @@ export function Control(props: ToolbarProps & { onClick: () => void }) {
         <Toolbar.Item
           key={item.id}
           id={item.id}
-          title={item.title}
+          tooltip={item.tooltip}
           onClick={() => {
             setActive(item.id);
             onClick();
@@ -218,7 +218,7 @@ export function WithTag() {
         <Toolbar.Item
           key={item.id}
           id={item.id}
-          title={item.title}
+          tooltip={item.tooltip}
           onClick={() => {
             setActive(item.id);
           }}
@@ -243,8 +243,8 @@ export function WithTag() {
 }
 export function Vertical() {
   const [state, setState] = useState(itemsBlueprintIcons[1]);
-  function handleChange({ id, icon, title }: ToolbarItemProps) {
-    setState({ id: id as string, icon, title });
+  function handleChange({ id, icon, tooltip }: ToolbarItemProps) {
+    setState({ id: id as string, icon, tooltip });
   }
 
   return (
@@ -262,17 +262,17 @@ export function Vertical() {
           <Toolbar.Item
             key={item.id}
             id={item.id}
-            title={item.title}
-            active={state.title === item.title}
+            tooltip={item.tooltip}
+            active={state.tooltip === item.tooltip}
             onClick={handleChange}
             icon={item.icon}
           />
         ))}
-        <Toolbar.Item title="Inbox" icon="inbox" />
+        <Toolbar.Item tooltip="Inbox" icon="inbox" />
       </Toolbar>
       <div style={{ padding: 5 }}>
         <p>Hello, World!</p>
-        <p>Value selected: {state.title}</p>
+        <p>Value selected: {state.tooltip}</p>
       </div>
     </div>
   );
@@ -281,8 +281,8 @@ export function Vertical() {
 export function Horizontal() {
   const [state, setState] = useState(itemsBlueprintIcons[1]);
 
-  function handleChange({ id, icon, title }: ToolbarItemProps) {
-    setState({ id: id as string, icon, title });
+  function handleChange({ id, icon, tooltip }: ToolbarItemProps) {
+    setState({ id: id as string, icon, tooltip });
   }
 
   return (
@@ -292,7 +292,7 @@ export function Horizontal() {
           <Toolbar.Item
             key={item.id}
             id={item.id}
-            title={item.title}
+            tooltip={item.tooltip}
             active={state.id === item.id}
             onClick={handleChange}
             icon={item.icon}
@@ -303,7 +303,7 @@ export function Horizontal() {
       </Toolbar>
       <div style={{ padding: 5 }}>
         <p>Hello, World!</p>
-        <p>Value selected: {state.title}</p>
+        <p>Value selected: {state.tooltip}</p>
       </div>
     </div>
   );
@@ -378,7 +378,7 @@ export function MixedItems(
           <Toolbar.Item
             key={itemProps.id}
             id={itemProps.id}
-            title={itemProps.title}
+            tooltip={itemProps.tooltip}
             onClick={() => {
               setActive(itemProps.id);
               onClick();
@@ -401,3 +401,149 @@ MixedItems.argTypes = {
     control: { type: 'radio' },
   },
 };
+
+interface CustomTipItem {
+  // eslint-disable-next-line react/no-unused-prop-types
+  id: string;
+  // eslint-disable-next-line react/no-unused-prop-types
+  icon: ToolbarItemProps['icon'];
+  title: string;
+  shortcuts: string[];
+  description: string;
+}
+
+const itemsShortcuts: CustomTipItem[] = [
+  {
+    id: 'copy',
+    icon: 'phone',
+    title: 'Copy text tool',
+    shortcuts: ['Ctrl', 'C'],
+    description: 'Copy selected item to clipboard.',
+  },
+  {
+    id: 'paste',
+    icon: 'add-column-left',
+    title: 'Paste text tool',
+    shortcuts: ['Ctrl', 'V'],
+    description: 'Paste copied item from clipboard.',
+  },
+  {
+    id: 'undo',
+    icon: 'add-column-right',
+    title: 'Undo text tool',
+    shortcuts: ['Ctrl', 'Z'],
+    description: 'Undo the last action.',
+  },
+  {
+    id: 'redo',
+    icon: 'redo',
+    title: 'Redo text tool',
+    shortcuts: ['Ctrl', 'Y'],
+    description: 'Redo the previously undone action.',
+  },
+  {
+    id: 'cut',
+    icon: 'cut',
+    title: 'Cut text tool',
+    shortcuts: ['Ctrl', 'X'],
+    description: 'Cut selected item to clipboard.',
+  },
+];
+
+export function CustomTipContent() {
+  return (
+    <div
+      style={{
+        display: 'flex',
+        height: 200,
+      }}
+    >
+      <Toolbar vertical>
+        {itemsShortcuts.map((item) => (
+          <Toolbar.Item
+            key={item.id}
+            id={item.id}
+            tooltip={<CustomTip {...item} />}
+            icon={item.icon}
+            tooltipProps={{ minimal: true }}
+          />
+        ))}
+        <Toolbar.Item tooltip="Inbox" icon="inbox" />
+      </Toolbar>
+    </div>
+  );
+}
+
+const styles: Record<
+  | 'titleContainer'
+  | 'title'
+  | 'description'
+  | 'shortcutContainer'
+  | 'shortcutItem',
+  CSSProperties
+> = {
+  titleContainer: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    flexWrap: 'wrap',
+    alignItems: 'center',
+  },
+  title: {
+    fontSize: '0.9rem',
+    minWidth: '50%',
+    padding: '5px 0',
+  },
+  description: {
+    paddingTop: '3rem',
+    fontSize: '0.7rem',
+  },
+  shortcutContainer: {
+    display: 'flex',
+    textWrap: 'nowrap',
+  },
+  shortcutItem: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: '0.5em',
+    border: '1px solid #ccc',
+    borderRadius: '4px',
+    height: 'calc(0.5em + 15px)',
+    minWidth: 'calc(0.5em + 15px)' /* Adjust padding and border width */,
+    marginLeft: '5px',
+    fontSize: '0.6rem',
+    fontWeight: 'bolder',
+  },
+};
+
+function CustomTip({ title, shortcuts, description }: CustomTipItem) {
+  return (
+    <div style={{ width: 250, padding: '0.5rem' }}>
+      <div style={styles.titleContainer}>
+        <span style={styles.title}>{title}</span>
+        <div style={styles.shortcutContainer}>
+          {shortcuts?.map((key, index) => {
+            return (
+              <div
+                key={key}
+                style={{
+                  ...styles.shortcutItem,
+                  ...(index === 0 && { margin: 0 }),
+                }}
+              >
+                <span>{key}</span>{' '}
+              </div>
+            );
+          })}
+        </div>
+      </div>
+
+      <p style={styles.description}>
+        {description}{' '}
+        <a target="_blank" href="">
+          Learn more
+        </a>{' '}
+      </p>
+    </div>
+  );
+}